pytorch_version/requirements_pytorch.txt
# PyTorch PINN Requirements for Linear Elasticity
# Python 3.11.x compatible  
# Generated on 2024-09-30

# Core dependencies
contourpy==1.3.3
cycler==0.12.1
filelock==3.19.1
fonttools==4.60.1
fsspec==2025.9.0
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.6
mpmath==1.3.0
networkx==3.5
numpy==2.3.3
packaging==25.0
pillow==11.3.0
pyDOE3==1.4.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
scipy==1.16.2
six==1.17.0
sympy==1.14.0
torch==2.8.0
typing_extensions==4.15.0
scikit-learn==1.3.0
